Prevent taxonomy terms to be created if not found by migration_lookup
If migration_lookup does not find a related entity and the no_stub is not set
to true, it will create a stub for it. In this case, it would effectively create
a new taxonomy term when the intended behavior is to only make the assignment
if a term already exist in the system from a previous migration.